+++ src/SellFeed/Services/SellFeedService.php
@@ -36,6 +36,17 @@ class SellFeedService extends \DTS\eBaySDK\SellFeed\Services\SellFeedBaseService
                 ],
             ]
         ],
+        'getTask' => [
+            'method' => 'GET',
+            'resource' => 'task/{task_id}',
+            'responseClass' => '\DTS\eBaySDK\SellFeed\Types\Task',
+            'params' => [
+                'task_id' => [
+                    'valid' => ['string'],
+                    'required' => true
+                ],
+            ]
+        ],
         'getTasks' => [
             'method' => 'GET',
             'resource' => 'task',
@@ -76,6 +87,23 @@ class SellFeedService extends \DTS\eBaySDK\SellFeed\Services\SellFeedBaseService
                     'valid' => ['string'],
                     'required' => true
                 ],
+            ]
+        ],
+        'createInventoryTask' => [
+            'method' => 'POST',
+            'resource' => 'inventory_task',
+            'responseClass' => '\DTS\eBaySDK\SellFeed\Types\CreateInventoryTaskResponse',
+            'params' => []
+        ],
+        'getInventoryTask' => [
+            'method' => 'GET',
+            'resource' => 'inventory_task/{task_id}',
+            'responseClass' => '\DTS\eBaySDK\SellFeed\Types\GetInventoryTaskResponse',
+            'params' => [
+                'task_id' => [
+                    'valid' => ['string'],
+                    'required' => true
+                ]
             ]
         ],
     ];
@@ -125,6 +153,24 @@ class SellFeedService extends \DTS\eBaySDK\SellFeed\Services\SellFeedBaseService
     }
 
     /**
+     * @param \DTS\eBaySDK\SellFeed\Types\GetTaskRequest $request
+     * @return \DTS\eBaySDK\SellFeed\Types\Task
+     */
+    public function getTask(\DTS\eBaySDK\SellFeed\Types\GetTaskRequest $request)
+    {
+        return $this->getTaskAsync($request)->wait();
+    }
+
+    /**
+     * @param \DTS\eBaySDK\SellFeed\Types\GetTaskRequest $request
+     * @return \GuzzleHttp\Promise\PromiseInterface
+     */
+    public function getTaskAsync(\DTS\eBaySDK\SellFeed\Types\GetTaskRequest $request)
+    {
+        return $this->callOperationAsync('getTask', $request);
+    }
+
+    /**
      * @param \DTS\eBaySDK\SellFeed\Types\GetTasksRequest $request
      * @return \DTS\eBaySDK\SellFeed\Types\GetTasksResponse
      */
@@ -158,5 +204,38 @@ class SellFeedService extends \DTS\eBaySDK\SellFeed\Services\SellFeedBaseService
     public function uploadFileAsync(\DTS\eBaySDK\SellFeed\Types\UploadFileRequest $request)
     {
         return $this->callFormOperationAsync('uploadFile', $request);
+    }
+
+    /**
+     * @param \DTS\eBaySDK\SellFeed\Types\CreateInventoryTaskRequest $request
+     * @return \DTS\eBaySDK\SellFeed\Types\CreateInventoryTaskResponse
+     */
+    public function createInventoryTask(\DTS\eBaySDK\SellFeed\Types\CreateInventoryTaskRequest $request)
+    {
+        return $this->createInventoryTaskAsync($request)->wait();
+    }
+    /**
+     * @param \DTS\eBaySDK\SellFeed\Types\CreateInventoryTaskRequest $request
+     * @return \GuzzleHttp\Promise\PromiseInterface
+     */
+    public function createInventoryTaskAsync(\DTS\eBaySDK\SellFeed\Types\CreateInventoryTaskRequest $request)
+    {
+        return $this->callOperationAsync('createInventoryTask', $request);
+    }
+    /**
+     * @param \DTS\eBaySDK\SellFeed\Types\GetInventoryTaskRequest $request
+     * @return \DTS\eBaySDK\SellFeed\Types\GetInventoryTaskResponse
+     */
+    public function getInventoryTask(\DTS\eBaySDK\SellFeed\Types\GetInventoryTaskRequest $request)
+    {
+        return $this->getInventoryTaskAsync($request)->wait();
+    }
+    /**
+     * @param \DTS\eBaySDK\SellFeed\Types\GetInventoryTaskRequest $request
+     * @return \GuzzleHttp\Promise\PromiseInterface
+     */
+    public function getInventoryTaskAsync(\DTS\eBaySDK\SellFeed\Types\GetInventoryTaskRequest $request)
+    {
+        return $this->callOperationAsync('getInventoryTask', $request);
     }
 }