+++ src/SellFeed/Types/GetTaskRequest.php
@@ -8,17 +8,24 @@
  * is updated. You've been warned!
  */
 
-namespace DTS\eBaySDK\PostOrder\Types;
+namespace DTS\eBaySDK\SellFeed\Types;
 
 /**
  *
+ * @property string $task_id
  */
-class ReturnShippingCostType extends \DTS\eBaySDK\Types\BaseType
+class GetTaskRequest extends \DTS\eBaySDK\Types\BaseType
 {
     /**
      * @var array Properties belonging to objects of this class.
      */
     private static $propertyTypes = [
+        'task_id' => [
+            'type' => 'string',
+            'repeatable' => false,
+            'attribute' => false,
+            'elementName' => 'task_id'
+        ],
     ];
 
     /**