+++ src/SellFeed/Types/InventoryFilterCriteria.php
@@ -8,46 +8,47 @@
  * is updated. You've been warned!
  */
 
-namespace DTS\eBaySDK\PostOrder\Types;
+namespace DTS\eBaySDK\SellFeed\Types;
 
 /**
  *
- * @property \DTS\eBaySDK\PostOrder\Types\Amount $amount
- * @property \DTS\eBaySDK\PostOrder\Enums\ReturnShippingCostTypeEnum $returnShippingCostType
+ * @property \DTS\eBaySDK\SellFeed\Enums\ListingFormatEnum $listingFormat
+ * @property string[] $listingStatuses
+ * @property string $dateRange
  */
-class ItemizedReturnShippingCostType extends \DTS\eBaySDK\Types\BaseType
+class InventoryFilterCriteria extends \DTS\eBaySDK\Types\BaseType
 {
     /**
      * @var array Properties belonging to objects of this class.
      */
     private static $propertyTypes = [
-        'amount' => [
-            'type' => 'DTS\eBaySDK\PostOrder\Types\Amount',
-            'repeatable' => false,
-            'attribute' => false,
-            'elementName' => 'amount'
-        ],
-        'returnShippingCostType' => [
+        'listingFormat' => [
             'type' => 'string',
             'repeatable' => false,
             'attribute' => false,
-            'elementName' => 'returnShippingCostType'
+            'elementName' => 'listingFormat'
+        ],
+        'listingStatuses' => [
+            'type' => 'string',
+            'repeatable' => true,
+            'attribute' => false,
+            'elementName' => 'listingStatuses'
+        ],
+        'dateRange' => [
+            'type' => 'string',
+            'repeatable' => false,
+            'attribute' => false,
+            'elementName' => 'dateRange'
         ]
     ];
 
-    /**
-     * @param array $values Optional properties and values to assign to the object.
-     */
     public function __construct(array $values = [])
     {
         list($parentValues, $childValues) = self::getParentValues(self::$propertyTypes, $values);
-
         parent::__construct($parentValues);
-
         if (!array_key_exists(__CLASS__, self::$properties)) {
             self::$properties[__CLASS__] = array_merge(self::$properties[get_parent_class()], self::$propertyTypes);
         }
-
         $this->setValues(__CLASS__, $childValues);
     }
 }